Code Quality: Introducing StandardViewBase to reduce code duplication (#10996)
