+++ src/Files.App/App.xaml
@@ -37,7 +37,7 @@
 					<GradientStop Offset="1.0" Color="{ThemeResource ControlStrokeColorDefault}" />
 				</LinearGradientBrush.GradientStops>
 			</LinearGradientBrush>
-			
+
 			<ResourceDictionary.MergedDictionaries>
 				<XamlControlsResources xmlns="using:Microsoft.UI.Xaml.Controls" />
 				<ResourceDictionary>
