+++ src/Files.App/Dialogs/CreateShortcutDialog.xaml
@@ -7,8 +7,8 @@
 	xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
 	Title="{helpers:ResourceString Name=NewShortcutDialogTitle}"
 	DefaultButton="Primary"
-	PrimaryButtonCommand="{x:Bind ViewModel.PrimaryButtonCommand}"
 	IsPrimaryButtonEnabled="{x:Bind ViewModel.IsLocationValid, Mode=OneWay}"
+	PrimaryButtonCommand="{x:Bind ViewModel.PrimaryButtonCommand}"
 	PrimaryButtonText="{helpers:ResourceString Name=Create}"
 	RequestedTheme="{x:Bind helpers:ThemeHelper.RootTheme}"
 	SecondaryButtonText="{helpers:ResourceString Name=Cancel}"
@@ -29,11 +29,11 @@
 				<RowDefinition Height="Auto" />
 				<RowDefinition Height="Auto" />
 			</Grid.RowDefinitions>
-			
+
 			<!--  Header  -->
-			<TextBlock 
+			<TextBlock
 				Grid.ColumnSpan="2"
-				Margin="0, 0, 0, 20"
+				Margin="0,0,0,20"
 				Text="{helpers:ResourceString Name=NewShortcutDialogDescription}"
 				TextWrapping="Wrap" />
