+++ src/Files.App/UserControls/FilePreviews/CodePreview.xaml
@@ -11,9 +11,7 @@
 	Loaded="UserControl_Loaded"
 	mc:Ignorable="d">
 	<Border>
-		<ScrollViewer
-			HorizontalScrollBarVisibility="Auto"
-			HorizontalScrollMode="Auto">
+		<ScrollViewer HorizontalScrollBarVisibility="Auto" HorizontalScrollMode="Auto">
 			<RichTextBlock
 				Name="codeView"
 				Padding="10"
