+++ src/Files.App/UserControls/FilePreviews/MediaPreview.xaml
@@ -24,9 +24,7 @@
 			Foreground="Transparent"
 			Source="{x:Bind ViewModel.Source, Mode=OneWay}">
 			<MediaPlayerElement.TransportControls>
-				<MediaTransportControls
-					IsPlaybackRateButtonVisible="False"
-					IsZoomButtonVisible="False" />
+				<MediaTransportControls IsPlaybackRateButtonVisible="False" IsZoomButtonVisible="False" />
 			</MediaPlayerElement.TransportControls>
 		</MediaPlayerElement>
 	</Border>
