+++ src/Files.App/UserControls/SidebarControl.xaml
@@ -42,13 +42,13 @@
 		<ResourceDictionary>
 			<!--  Removes NavigationView background  -->
 			<SolidColorBrush x:Key="NavigationViewContentBackground" Color="Transparent" />
-			
+
 			<!--  Removes NavigationView corner radius  -->
 			<CornerRadius x:Key="NavigationViewContentGridCornerRadius">0</CornerRadius>
 
 			<!--  Removes NavigationView border from top, right and bottom  -->
 			<Thickness x:Key="NavigationViewContentGridBorderThickness">1,0,0,0</Thickness>
-			
+
 			<mconv:BoolNegationConverter x:Key="BoolNegationConverter" />
 			<mconv:StringFormatConverter x:Key="StringFormatConverter" />
 			<vc:ColorModelToColorConverter x:Key="ColorModelToColorConverter" />
@@ -420,8 +420,8 @@
 								<Grid
 									x:Name="ContentRoot"
 									Grid.Row="2"
-									Background="{ThemeResource App.Theme.FileArea.BackgroundBrush}"
-									Grid.ColumnSpan="2">
+									Grid.ColumnSpan="2"
+									Background="{ThemeResource App.Theme.FileArea.BackgroundBrush}">
 									<Grid.RenderTransform>
 										<CompositeTransform x:Name="ContentTransform" />
 									</Grid.RenderTransform>
@@ -1195,6 +1195,6 @@
 			DriveNavItemTemplate="{StaticResource DriveNavItem}"
 			FileTagNavItemTemplate="{StaticResource FileTagNavItem}"
 			LinuxNavItemTemplate="{StaticResource LinuxNavItem}"
-			LocationNavItemTemplate="{StaticResource LocationNavItem}"/>
+			LocationNavItemTemplate="{StaticResource LocationNavItem}" />
 	</NavigationView.MenuItemTemplateSelector>
 </NavigationView>
