+++ src/Files.App/Views/LayoutModes/ColumnViewBase.xaml
@@ -1,4 +1,4 @@
-﻿<local:BaseLayout
+﻿<local:StandardViewBase
 	x:Class="Files.App.Views.LayoutModes.ColumnViewBase"
 	xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
 	xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
@@ -15,7 +15,7 @@
 	xmlns:vc="using:Files.App.ValueConverters"
 	x:Name="PageRoot"
 	mc:Ignorable="d">
-	<local:BaseLayout.Resources>
+	<local:StandardViewBase.Resources>
 		<vc:ColorModelToColorConverter x:Key="ColorModelToColorConverter" />
 		<converters:BoolNegationConverter x:Key="BoolNegationConverter" />
 		<converters:BoolToVisibilityConverter
@@ -120,7 +120,7 @@
 				</Setter.Value>
 			</Setter>
 		</Style>
-	</local:BaseLayout.Resources>
+	</local:StandardViewBase.Resources>
 
 	<Grid x:Name="RootGrid" ContextFlyout="{x:Bind BaseContextMenuFlyout}">
 		<uc:FolderEmptyIndicator
@@ -449,4 +449,4 @@
 				StrokeThickness="1" />
 		</Canvas>
 	</Grid>
-</local:BaseLayout>
+</local:StandardViewBase>
