+++ src/Files.App/Views/LayoutModes/DetailsLayoutBrowser.xaml
@@ -1,4 +1,4 @@
-<local:BaseLayout
+<local:StandardViewBase
 	x:Class="Files.App.Views.LayoutModes.DetailsLayoutBrowser"
 	xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
 	xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
@@ -28,7 +28,7 @@
 			<icore:InvokeCommandAction Command="{x:Bind CommandsViewModel.ItemPointerPressedCommand}" />
 		</icore:EventTriggerBehavior>
 	</i:Interaction.Behaviors>
-	<local:BaseLayout.Resources>
+	<local:StandardViewBase.Resources>
 		<ResourceDictionary>
 			<ResourceDictionary.MergedDictionaries>
 				<ResourceDictionary Source="ms-appx:///ResourceDictionaries/DefaultGridSplitterStyle.xaml" />
@@ -163,7 +163,7 @@
 				<Setter Property="Opacity" Value="0.6" />
 			</Style>
 		</ResourceDictionary>
-	</local:BaseLayout.Resources>
+	</local:StandardViewBase.Resources>
 
 	<Grid
 		x:Name="RootGrid"
@@ -1010,4 +1010,4 @@
 			</VisualStateGroup>
 		</VisualStateManager.VisualStateGroups>
 	</Grid>
-</local:BaseLayout>
+</local:StandardViewBase>
