+++ src/Files.App/Views/LayoutModes/GridViewBrowser.xaml
@@ -1,4 +1,4 @@
-<local:BaseLayout
+<local:StandardViewBase
 	x:Class="Files.App.Views.LayoutModes.GridViewBrowser"
 	xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
 	xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
@@ -29,7 +29,7 @@
 			<icore:InvokeCommandAction Command="{x:Bind CommandsViewModel.ItemPointerPressedCommand}" />
 		</icore:EventTriggerBehavior>
 	</i:Interaction.Behaviors>
-	<local:BaseLayout.Resources>
+	<local:StandardViewBase.Resources>
 		<vc:ColorModelToColorConverter x:Key="ColorModelToColorConverter" />
 		<converters:BoolNegationConverter x:Key="BoolNegationConverter" />
 		<converters:BoolToVisibilityConverter
@@ -474,7 +474,7 @@
 				</Setter.Value>
 			</Setter>
 		</Style>
-	</local:BaseLayout.Resources>
+	</local:StandardViewBase.Resources>
 
 	<Grid
 		x:Name="RootGrid"
@@ -736,4 +736,4 @@
 				StrokeThickness="1" />
 		</Canvas>
 	</Grid>
-</local:BaseLayout>
+</local:StandardViewBase>
