+++ src/Files.App/Views/Pages/Properties.xaml
@@ -19,10 +19,10 @@
 				<Thickness x:Key="NavigationViewTopPaneHeight">40</Thickness>
 				<Thickness x:Key="TopNavigationViewTopNavGridMargin">0</Thickness>
 				<SolidColorBrush x:Key="NavigationViewContentGridBorderBrush" Color="Transparent" />
-				
+
 				<!--  Removes NavigationView background  -->
 				<SolidColorBrush x:Key="NavigationViewContentBackground" Color="Transparent" />
-				
+
 				<!--  Removes NavigationView corner radius  -->
 				<CornerRadius x:Key="NavigationViewContentGridCornerRadius">0</CornerRadius>
 			</ResourceDictionary>
