+++ src/Files.App/Views/Pages/PropertiesShortcut.xaml
@@ -128,11 +128,11 @@
 				Text="{x:Bind ViewModel.ShortcutItemWorkingDir, Mode=TwoWay}" />
 		</Grid>
 
-        <local1:SettingsBlockControl 
-	        x:Name="RunAsAdminCheckbox" 
-	        Title="{helpers:ResourceString Name=RunAsAdministrator}" 
-	        HorizontalAlignment="Stretch" 
-	        x:Load="{x:Bind ViewModel.RunAsAdminEnabled, Mode=OneWay}">
+		<local1:SettingsBlockControl
+			x:Name="RunAsAdminCheckbox"
+			Title="{helpers:ResourceString Name=RunAsAdministrator}"
+			HorizontalAlignment="Stretch"
+			x:Load="{x:Bind ViewModel.RunAsAdminEnabled, Mode=OneWay}">
 			<ToggleSwitch
 				AutomationProperties.Name="{helpers:ResourceString Name=RunAsAdministrator}"
 				IsOn="{x:Bind ViewModel.RunAsAdmin, Mode=TwoWay}"
